+++ polugins_type_gen/tests/pkgs/example_package/src/example_package/same_module/some_file_2.py
@@ -1,0 +1,10 @@
+import polars as pl
+
+
+class NestedSameModule2PackageNamespace:
+    def __init__(self, ldf: pl.LazyFrame):
+        self._ldf = ldf
+
+    def some_method(self, x: int) -> pl.LazyFrame:
+        """my docstring"""
+        return self._ldf