+++ polugins_type_gen/tests/test_cli.py
@@ -1,4 +1,3 @@
-import shutil
 from pathlib import Path
 
 import pytest
@@ -12,24 +11,27 @@ from polugins_type_gen.cli import (
 typings_dir = Path("typings")
 
 
-@pytest.fixture
-def typings_directory():
-    if typings_dir.exists():
-        shutil.rmtree(typings_dir)
-    yield
-    if typings_dir.exists():
-        shutil.rmtree(typings_dir)
-
-
-@pytest.mark.skip()
-def test_cli(typings_directory):
+def test_cli():
     create_stubs()
     lazyframe_stubs_path = (typings_dir / ExtensionClass.LAZYFRAME.import_path).with_suffix(".pyi")
     lazyframe_stubs = lazyframe_stubs_path.read_text()
     assert "external: example_package.PackageNamespace" in lazyframe_stubs
+    assert "external_nested: example_package.some_module.some_file.NestedPackageNamespace"
+    assert "external_nested_init: example_package.some_module_from_init.NestedInitPackageNamespace"
+    assert "external_nested_same_module_1: example_package.same_module.some_file_1.NestedSameModule1PackageNamespace"
+    assert "external_nested_same_module_2: example_package.same_module.some_file_2.NestedSameModule2PackageNamespace"
     assert "pyproject: polugins._testing.namespaces.PyProjectNameSpace" in lazyframe_stubs
     assert "config: polugins._testing.namespaces.ConfigNameSpace" in lazyframe_stubs
     assert "env: polugins._testing.namespaces.EnvNameSpace" in lazyframe_stubs
+
+    assert "import example_package" in lazyframe_stubs
+    assert "import example_package.some_module.some_file"
+    assert "import example_package.some_module_from_init"
+    assert "import example_package.same_module.some_file_1"
+    assert "import example_package.same_module.some_file_2"
+    assert "import polugins._testing.namespaces" in lazyframe_stubs
+    assert "import polugins._testing.namespaces" in lazyframe_stubs
+    assert "import polugins._testing.namespaces" in lazyframe_stubs
 
 
 def test_exception_on_no_namespaces(mocker):